Fix #5282: Support mixed media tweets.
Switch to using the `https://cdn.syndication.twimg.com/tweet-result?id=$status_id` API for Twitter.
This API is almost the same as the regular v1 API, except it returns all media attachments for the tweet.

The syndication API is used by Twitter's embedded tweet widget when a tweet is embedded in another
website. So it's an internal API, but it's probably safe for us to use. It doesn't require
authentication, even for NSFW tweets.

* https://publish.twitter.com/?query=https%3A%2F%2Ftwitter.com%2Ftwotenky%2Fstatus%2F1577831592227000320&widget=Tweet
* https://cdn.syndication.twimg.com/tweet-result?id=1577831592227000320
* https://twitter.com/twotenky/status/1577831592227000320 (click "..." then click "Embed Tweet")

Closes #4656: Migrate to Twitter v2 API