posts: remove "save this video" links.
Remove the "Save this video (right click and save)" links beneath video and ugoira posts. These
aren't necessary since you can simply right click and save the video itself, or use the Download
link in the sidebar.

Also fix it so that when viewing a ugoira post, you can switch back to the .webm sample. Also fix it
so the URL doesn't lose the URL params when switching between the .webm sample and the original ugoira.