Fix #5399: some skeb page(s) fail to upload
Fix a bug where certain .webp files failed to upload because they weren't detected as .webp. This
happened when the filesize in the .webp header contained a 0x0A byte. This was because the file type
is detected using a regex, and the 0x0A byte broke the regex because it was interpreted as a
newline, which the regex didn't match because regexes in Ruby don't match newlines by default.

Add a test for this, but use a truncated .webp instead of the full file to save space.