Upgrade to Ruby 3.2.0.
Upgrade from Ruby 3.1.2 to 3.2.0.

The headline feature is JIT compilation with YJIT. YJIT is enabled by default in the Docker image.
Outside of Docker, you must enable it manually by setting the RUBY_YJIT_ENABLE=1 environment
variable. This variable must be set before Ruby starts, so setting it in .env won't work.

This adds an optional build-time dependency on Rust to enable YJIT.

To test that YJIT is working, do `ruby --version` and look for `+YJIT`, or do
`RubyVM::YJIT.enabled?` in `bin/rails console`.

https://www.ruby-lang.org/en/news/2022/12/25/ruby-3-2-0-released/
https://github.com/ruby/ruby/blob/master/doc/yjit/yjit.md