docker: refactor Dockerfile for faster builds.
Refactor the Dockerfile to improve build speed when a base dependency is updated. Split the base
image up into multiple stages so that when one base dependency changes (Ruby, FFmpeg, Vips, etc), it
doesn't require rebuilding everything else.

This also allows base dependencies to be built in parallel. This is disabled in CI because building
multiple layers in parallel may run out of memory when building the ARM image in QEMU.