Fix danbooru/danbooru#5398: Unclosed [b] causes [/quote] to not work properly
Fix the [/quote] tag to close all open tags up to the last [quote].

This fixes about 300 messages with unclosed tags inside a [quote] block.

There's one known bug: unclosed [code] tags inside a [quote] aren't closed until they find a [/code]
tag or the end of the message. So if you have this:

    [quote][code]foo bar[/quote]

Then it will be treated like this:

    <blockquote><pre>foo bar[/quote]</pre></blockquote>

Not like this:

    <blockquote><pre>foo bar</pre></blockquote>

Fixing this would require some sort of lookahead or multiple passes to know when an open [code] tag
is never closed.