Fix r888888888/dtext_rb#38: Add an ability to use <hr> tags
Add a `[hr]` (horizontal rule) tag for generating horizontal lines.

The tag can be written as either `[hr]` or `<hr>`. The tag must appear on a line by itself. It can't
appear in the middle of a paragraph, header, or list.

Known bug: `[quote][hr]\nblah\n[/quote]` works when it shouldn't. A newline is supposed to be
required between the [quote] and [hr] tags, but it works without one.

Known bug: Putting a [hr] tag inside a block [tn] tag, as in `[tn]\n[hr]\n[/tn]`, generates the invalid
HTML `<p class="tn"><hr></p>`. This is invalid because the <p> tag can't contain the <hr> tag.

Known bug: Markdown-style syntax, where you can write `---` on a line by itself to make a <hr> tag,
isn't supported. This is because there are about 20k commentaries using this syntax, and it's hard
to tell if parsing them as [hr] tags would break things.