Fix lowercasing of Unicode wiki links.
Fix it so that wiki links are normalized to ASCII lowercase, not Unicode lowercase.

For example, fix it so that the link `[[Full Metal Panic! Σ]]` is normalized to
`/wiki_pages/full_metal_panic_Σ`, not `/wiki_pages/full_metal_panic_ς`. Neither link actually works
because Unicode is not allowed in tag names, but even so, lowercasing Unicode characters is not
correct behavior because it breaks the link. ASCII-only lowercasing is also a little bit faster.