Fix #5416: Encoding::CompatibilityError on /status
Fix encoding compatibility error that happened when we tried to display the request's HTTP headers
on the status page.

HTTP headers are encoded as ASCII-8BIT strings, and when an ASCII-8BIT string is concatenated with a
UTF-8 string (as when it's rendered in the HTML page), it will raise an error if the string contains
any non-ASCII bytes. The solution is to change the encoding from ASCII-8BIT to UTF-8. But note that
we can't just do `force_encoding("UTF-8")`, because that will raise a different error if the header
contains invalid UTF-8 characters.