Use consistent convention for end of string pointers.
We frequently deal with `a` and `b` pointers denoting the start and end of a string, but previously
we weren't consistent on whether the end pointer pointed to the last character, or one past the last
character. Fix it so that end pointers always point to one past the last character. Using a
consistent convention eliminates a lot of +1 and -1 offset calculations, plus it makes it easier to
wrap pointers in std::string_view objects so we can treat them like strings.