Fix lists incorrectly closing previous tags.
Fix a bug where lists incorrectly closed previous tags at the end of the list.

This commonly happened when two lists were separated by a header with no blank lines in between. For
example, this:

    h4. See also
    * list one
    h4. External links
    * list two

Produced this:

    <h4>See also</h4>
    <ul>
      <li>One</li>
    </ul>
    <h4>External links</h4>
    <li>Two</li>

Here the parser didn't recognize the first list was closed, so it thought the second list was still
part of the first list and generated invalid HTML. This appeared to work because browsers are
tolerant of <li> tags outside of <ul> tags, even though it's incorrect.

Another common case was when a list was contained in two nested quotes. In this case the list would
incorrectly close its parent element immediately when the list ended.

For example, this:

    [quote]
    A said:

    [quote]
    B said:

    * One
    * Two
    [/quote]
    [/quote]

    Reply.

Produced this:

    <blockquote>
      <p>A said:</p>
      <blockquote>
        <p>B said:</p>
        <ul>
          <li>One</li>
          <li>Two<br></li>
        </ul>
      </blockquote>
    </blockquote>
    [/quote]Reply

Here the list incorrectly closed its parent tag (the inner [quote]) immediately when the list ended.
This caused the last `[/quote]` to be treated as an unexpected stray tag, because the [quote] tag
had already been incorrectly closed by the list.

Fixes https://github.com/danbooru/danbooru/issues/1964#issuecomment-321988696.