Fix it so that list items can't be split onto multiple lines.
For example, treat this:

    * one
    two
    * three
    four

The same as if you had written this:

    * one

    two

    * three

    four

Before the first would generate this HTML:

    <ul>
      <li>
        one
        <br>two
      </li>
      <li>
        three
        <br>four
      <li>
    </ul>

Now it generates this HTML:

    <ul>
      <li>one</li>
    </ul>
    <p>two</p>
    <ul>
      <li>three</li>
    </ul>
    <p>four</p>

This was because list items were treated as paragraphs, and paragraphs are allowed to be split onto
multiple lines. However, it was usually (but not always) an accident for a user to do this with
lists. Usually it was the result of not putting a blank line between a list and the next paragraph.

This also fixes several other bugs:

* Fixes lists being limited to 512 items (a consequence of processing lists recursively).
* Fixes open tags not being closed before entering a list, or after each list item.
* Fixes a bug where if a list contained unclosed tags, or it was closed by a [/spoiler] tag in the
  middle of the list, then subsequent list items (<li> elements) would appear outside of the list
  (<ul> tag).
* Fixes not being able to put headers immediately after a list, with no blank line in between.

This changes the rendering of about 2000 wiki page versions.

Fixes danbooru/danbooru#1964: Text near a list will be joined into the last bullet point.