Fix parsing of blank lines.
Fix it so that lines containing only spaces are considered blank lines. Before a line containing
only spaces wasn't considered blank. This often caused lines containing extra spaces at the
beginning or end of a message to be turned into empty paragraphs, like `<p> </p>`.

This also fixes it so that spaces at the end of lines are removed.