Fix [code] and [nodtext] blocks to strip leading/trailing newlines.
Fix it so that [code] and [nodtext] blocks strip the first and last newlines in the block.

For example, this:

   [code]
   puts "Hello world"
   [/code]

Produces this:

   <pre>puts "Hello world"</pre>

Instead of this

   <pre>
   puts "Hello world"
   </pre>

This makes the [code] tag consistent with code fences.

This also fixes a bug where we swallowed _all_ leading whitespace in [code] blocks, which could break
ASCII art. For example, this fixes the inability to Triforce:

   [code]
    ▲
   ▲ ▲
   [/code]

Produced:

   ▲
   ▲ ▲