Add ability to link to IDs on wiki pages (e.g. [[Touhou#See also]])
Add ability to link to HTML IDs on a wiki page.

The ID must start with a capital letter:

* [[touhou#See also]] (links to the "See also" section on the Touhou page)
* [[touhou#see also]] (links to the tag literally named "touhou#see_also")
* [[Evo Moment #37]] (links to the tag literally named "evo_moment_#37")
* [[htol#niq:_hotaru_no_nikki|Htol#Niq: Hotaru no Nikki]] (links to the tag literally named "htol#niq:_hotaru_no_nikki")

The '#' should come before the '|':

* [[touhou#Games|List of Touhou games]] (right)
* [[touhou|List of Touhou games#Games]] (wrong)

The ID is lowercased and spaces and underscores are converted to dashes:

* [[touhou#See also]] (links to "/wiki_pages/touhou#dtext-see-also")

This means headers should be written like this: `h4#see-also. See also`.

Fixes danbooru/danbooru/4449.