Add [colgroup] and [col] tags for tables.
Allow tables to use the [colgroup] and [col] tags. These allow setting the alignment of entire
columns at a time.

Example:

    [table]
      [colgroup]
        [col align="left"]
        [col align="right" span="2"]
      [/colgroup]
      [tr]
        [td]Column 1[/td]
        [td]Column 2[/td]
        [td]Column 3[/td]
      [/tr]
      [tr]
        [td]...[/td]
        [td]...[/td]
        [td]...[/td]
      [/tr]
    [/table]

This example sets the first column to left align, and the second and third columns to right align.

Note that the [col] tag does not have a closing [/col] tag.

Fixes danbooru/danbooru#4784: Add colspan and rowspan to Dtext [td]