fantia: fix API call failures caused by missing CSRF token.
The `https://fantia.jp/api/v1/posts/:post_id` endpoint now requires the X-CSRF-Token header to be
set, otherwise it returns a 403 response. We have to get the CSRF token from the HTML page. The HTML
page doesn't contain the actual data for the post, so we still have to use the API even though we're
fetching the HTML page.