robots.txt: fix `source:` searches not being properly excluded.
Fix a bug where searches like this weren't properly excluded by robots.txt:

    https://danbooru.donmai.us/posts?page=71&tags=source:https://i.pximg.net/img-original/img/2020/02/23/11/10/40/79671797_p2.png
    https://danbooru.donmai.us/posts?page=17&tags=source:https://i.pximg.net/img-original/img/2019/09/14/21/38/33/76780395_p0.jpg
    https://danbooru.donmai.us/posts?page=4&tags=source:http://i1.pixiv.net/img-original/img/2016/01/21/00/05/49/54820200_p1.jpg

This lead to Googlebot spending lots of time crawling extremely slow `source:` searches.

We used to link to `source:` searches like this on the post page, but we removed these links long
ago and marked them as noindex, but for some reason Google is still hellbent on crawling them.