MediaFile: add pixel_hash method for detecting pixel-perfect duplicates (#4428)
Add a method for computing a hash of an image's pixel data. This doesn't add a way to store it on
media assets yet.

The algorithm is as follows:

* Return the file's MD5 if the image is animated or corrupt.
* Rotate the image if it has the EXIF orientation flag.
* Transform the image to sRGB if it has an embedded color profile.
* Transform the image to sRGB if it's still not sRGB (for example, it's 1-channel greyscale or 4-channel CMYK).
* Add an opaque alpha channel if the image doesn't have an alpha channel already.
* Write the image in .pam format [1].
* Compute the MD5 of the .pam file.

The .pam file format is a simple raw image format that consists of a plaintext header followed by
the raw pixel data. This format was chosen because it's easy to generate, it's readable by libvips
and ImageMagick, and it's viewable using the ImageMagick command `display file.pam`. This makes it
possible to visually inspect the image in case there are any problems with the hash.

This all assumes that decompressing an image is fully deterministic and the hash will always be the
same for different machines and different software. This may not be true. In particular:

* If there's any floating-point arithmetic involved, then rounding errors could cause the hash to
  change. This could occur when converting the image to sRGB from another colorspace.

* libjpeg has three different ways of calculating the discrete-cosine transform (DCT), each of which
  produces different results. See J_DCT_METHOD in libjpeg.txt[2]. There's a floating-point method, a
  fast-but-inaccurate integer method (JDCT_IFAST), and a slow-but-accurate integer method (JDCT_ISLOW).
  They can be compared with `djpeg -dct int file.jpg | md5sum`. JDCT_ISLOW is the method we use.

* libjpeg has two methods for upsampling chroma components during decompression, which produce
  different results. Compare `djpeg -nosmooth` with `djpeg`.

* libjpeg's djpeg produces different results compared to mozjpeg's and libjpeg-turbo's djpeg.

Partial fix for #4428: Add some sort of pixel-identical check to uploads.

[1]: https://netpbm.sourceforge.net/doc/pam.html
[2]: https://github.com/mozilla/mozjpeg/blob/master/libjpeg.txt