Fix danbooru#5436: GitHub-style dtext code blocks aren't rendered correctly in quotes
The bug was that when a code fence was terminated with "\r\n```" instead of "\n```" (Windows-style
line endings instead of Unix-style), it wasn't recognized as closed. The code block kept going until
the next [/quote] tag or the end of the message, whichever came first.

This was caused by defining newlines as `"\r\n" | "\n"` in Ragel. For some reason this didn't work
properly when used like `any* :>> newline "```"` in the code fence definition.

The easiest fix is to just convert "\r\n"-style line endings to "\n" upfront so we don't have to
worry about different line ending conventions everywhere.