dtext: fix undefined method error when code is reloaded.
Fix undefined method errors when a file was edited in development mode and code was reloaded. The
DText class didn't get properly reloaded because it's defined in two places, once in the gem and
once in app/logical/d_text.rb, and Zeitwerk doesn't like it when the same class is defined in two
different places.

Known bug: editing app/logical/d_text.rb doesn't reload the DText class to show the new changes.